refactor(service): consolidate cli files, follow bin naming convention
- Merge cli/index.ts + cli/main.ts → cli/cli.ts (no need for separate barrel)
- Rename bin/cli.ts → bin/sync-service.ts (matches bin name per package convention)
- Update package.json bin path accordingly

Committed-By-Agent: claude
