+++ apps/service/src/bin/sync-service.ts
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 import 'dotenv/config'
 import { runMain } from 'citty'
-import { createProgram } from '../cli/main.js'
+import { createProgram } from '../cli/cli.js'
 
 const program = await createProgram()
 runMain(program)
