--- apps/service/src/cli/index.ts
@@ -1,1 +1,0 @@
-export { createProgram } from './main.js'
